chore(cli): remove most calls to `commandDir`
