+++ packages/cli/src/commands/experimental.js
@@ -2,24 +2,29 @@ import terminalLink from 'terminal-link'
 
 import detectRwVersion from '../middleware/detectProjectRwVersion'
 
+import * as setupDockerCommand from './experimental/setupDocker'
+import * as setupInngestCommand from './experimental/setupInngest'
+import * as setupOpentelementryCommand from './experimental/setupOpentelemetry'
+import * as setupRscCommand from './experimental/setupRsc'
+import * as setupSentryCommand from './experimental/setupSentry'
+import * as setupServerFileCommand from './experimental/setupServerFile'
+import * as setupStreamingSsrCommand from './experimental/setupStreamingSsr'
+import * as studioCommand from './experimental/studio'
+
 export const command = 'experimental <command>'
 export const aliases = ['exp']
 export const description = 'Run or setup experimental features'
 
-export const builder = (yargs) =>
+export function builder(yargs) {
   yargs
-    .commandDir('./experimental', {
-      recurse: true,
-      // @NOTE This regex will ignore all commands nested more than two
-      // levels deep.
-      // e.g. /setup/hi.js & setup/hi/hi.js are picked up, but
-      // setup/hi/hello/bazinga.js will be ignored
-      // The [/\\] bit is for supporting both windows and unix style paths
-      // Also take care to not trip up on paths that have "setup" earlier
-      // in the path by eagerly matching in the start of the regexp
-      exclude: /.*[/\\]experimental[/\\].*[/\\].*[/\\]/,
-    })
-    .demandCommand()
+    .command(setupDockerCommand)
+    .command(setupInngestCommand)
+    .command(setupOpentelementryCommand)
+    .command(setupRscCommand)
+    .command(setupSentryCommand)
+    .command(setupServerFileCommand)
+    .command(setupStreamingSsrCommand)
+    .command(studioCommand)
     .middleware(detectRwVersion)
     .epilogue(
       `Also see the ${terminalLink(
@@ -27,3 +32,4 @@ export const builder = (yargs) =>
         'https://redwoodjs.com/docs/cli-commands#experimental'
       )}`
     )
+}
