+++ packages/cli/src/commands/record.js
@@ -1,15 +1,17 @@
-export const command = 'record <command>'
-export const description =
-  'Setup RedwoodRecord for your project. Caches a JSON version of your data model and adds api/src/models/index.js with some config.'
 import terminalLink from 'terminal-link'
 
-export const builder = (yargs) =>
+import * as initCommand from './record/init'
+
+export const command = 'record <command>'
+export const description = 'Set up RedwoodRecord for your project'
+
+export function builder(yargs) {
   yargs
-    .commandDir('./record', { recurse: false })
-    .demandCommand()
+    .command(initCommand)
     .epilogue(
       `Also see the ${terminalLink(
         'RedwoodRecord Docs',
         'https://redwoodjs.com/docs/redwoodrecord'
       )}\n`
     )
+}
