+++ packages/cli/src/commands/setup.js
@@ -2,23 +2,38 @@ import terminalLink from 'terminal-link'
 
 import detectRwVersion from '../middleware/detectProjectRwVersion'
 
+import * as authCommand from './setup/auth/auth'
+import * as cacheCommand from './setup/cache/cache'
+import * as customWebIndexCommand from './setup/custom-web-index/custom-web-index'
+import * as deployCommand from './setup/deploy/deploy'
+import * as generatorCommand from './setup/generator/generator'
+import * as graphiqlCommand from './setup/graphiql/graphiql'
+import * as i18nCommand from './setup/i18n/i18n'
+import * as mailerCommand from './setup/mailer/mailer'
+import * as packageCommand from './setup/package/package'
+import * as tsconfigCommand from './setup/tsconfig/tsconfig'
+import * as uiCommand from './setup/ui/ui'
+import * as viteCommand from './setup/vite/vite'
+import * as webpackCommand from './setup/webpack/webpack'
+
 export const command = 'setup <command>'
 export const description = 'Initialize project config and install packages'
 
-export const builder = (yargs) =>
+export function builder(yargs) {
   yargs
-    .commandDir('./setup', {
-      recurse: true,
-      // @NOTE This regex will ignore all commands nested more than two
-      // levels deep.
-      // e.g. /setup/hi.js & setup/hi/hi.js are picked up, but
-      // setup/hi/hello/bazinga.js will be ignored
-      // The [/\\] bit is for supporting both windows and unix style paths
-      // Also take care to not trip up on paths that have "setup" earlier
-      // in the path by eagerly matching in the start of the regexp
-      exclude: /.*[/\\]setup[/\\].*[/\\].*[/\\]/,
-    })
-    .demandCommand()
+    .command(authCommand)
+    .command(cacheCommand)
+    .command(customWebIndexCommand)
+    .command(deployCommand)
+    .command(generatorCommand)
+    .command(graphiqlCommand)
+    .command(i18nCommand)
+    .command(mailerCommand)
+    .command(packageCommand)
+    .command(tsconfigCommand)
+    .command(uiCommand)
+    .command(viteCommand)
+    .command(webpackCommand)
     .middleware(detectRwVersion)
     .epilogue(
       `Also see the ${terminalLink(
@@ -26,3 +41,4 @@ export const builder = (yargs) =>
         'https://redwoodjs.com/docs/cli-commands#setup'
       )}`
     )
+}
