+++ packages/cli/src/commands/setup/deploy/deploy.js
@@ -1,11 +1,23 @@
-export const command = 'deploy <target>'
-export const description = 'Setup deployment to various targets'
 import terminalLink from 'terminal-link'
 
-export const builder = (yargs) =>
+import * as baremetalCommand from './providers/baremetal'
+import * as coherenceCommand from './providers/coherence'
+import * as flightcontrolCommand from './providers/flightcontrol'
+import * as netlifyCommand from './providers/netlify'
+import * as renderCommand from './providers/render'
+import * as vercelCommand from './providers/vercel'
+
+export const command = 'deploy <target>'
+export const description = 'Setup deployment to various targets'
+
+export function builder(yargs) {
   yargs
-    .commandDir('./providers', { recurse: true })
-    .demandCommand()
+    .command(baremetalCommand)
+    .command(coherenceCommand)
+    .command(flightcontrolCommand)
+    .command(netlifyCommand)
+    .command(renderCommand)
+    .command(vercelCommand)
     .option('force', {
       alias: 'f',
       default: false,
@@ -18,3 +30,4 @@ export const builder = (yargs) =>
         'https://redwoodjs.com/docs/cli-commands#setup-deploy-config'
       )}`
     )
+}
